+++ src/hooks/useFamilyUpdate.ts
@@ -5,8 +5,8 @@ import {
 	transformFamilyDataBack,
 } from '@adapters/family.adapter';
 import { cleanNameURL } from './cleanNameUtil';
-import useErrorHandling from './useErrorHandling';
-import useSuccessHandling from './useSuccessHandling';
+import useErrorHandling from './common/useErrorHandling';
+import useSuccessHandling from './common/useSuccessHandling';
 
 interface ValidationError {
 	message: string[];
