+++ src/hooks/useFetchCategoriesFromFamily.ts
@@ -5,7 +5,7 @@ import {
 } from '@adapters/category.adapter';
 import axios from 'axios';
 import { useState } from 'react';
-import useErrorHandling from './useErrorHandling';
+import useErrorHandling from './common/useErrorHandling';
 
 function useFetchCategoriesFromFamily() {
 	const { errorLog, showError } = useErrorHandling();
