+++ src/interfaces/users/userResponse.ts
@@ -1,0 +1,10 @@
+import { UserDataBackend } from '@adapters/user.adapter';
+
+export interface RegisterResponse {
+	token: string;
+	user: UserDataBackend;
+}
+
+export interface ValidationError {
+	message: string[];
+}
