+++ src/pages/Login.tsx
@@ -1,7 +1,7 @@
 import { UserDataBackend, transformUserData } from '@adapters/user.adapter';
 import { AuthContext } from '@apps/App';
 import ErrorMessage from '@components/ErrorMessage';
-import useErrorHandling from '@hooks/useErrorHandling';
+import useErrorHandling from '@hooks/common/useErrorHandling';
 import axios, { AxiosError } from 'axios';
 import { jwtDecode } from 'jwt-decode';
 import { FormEvent, useContext, useEffect, useState } from 'react';
