+++ src/pages/userPages/UserOptionEdit.tsx
@@ -3,8 +3,8 @@ import ErrorMessage from '@components/ErrorMessage';
 import FileUpload from '@components/FileUpload';
 import HeaderTitle from '@components/HeaderTitle';
 import SuccessMessage from '@components/SuccessMessage';
-import usePasswordGenerator from '@hooks/usePasswordGenerator';
-import { useUserRoles } from '@hooks/useUserRoles';
+import usePasswordGenerator from '@hooks/common/usePasswordGenerator';
+import { useUserRoles } from '@hooks/userRoles/useUserRoles';
 import useUserUpdate from '@hooks/useUserUpdate';
 import { UserRole } from '@interfaces/UserRole';
 import BasePage from '@layouts/BasePage';
